Project restructuring with separate controllers and utils directories
* routes restructured into separate directories

* imports updated

* home page controller added

* anime streaming routes and controllers updated

* content loader routes and controllers updated

* search routes and controllers updated

* manga routes and controllers updated
