+++ routes/loader/loader.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const { LoaderController } = require("../../controllers/loader.controller");
+
+const router = express.Router();
+
+router.post("/", LoaderController);
+
+module.exports = router;
