+++ routes/loader/loadpage.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const { RenderLoadEpisodes } = require("../../controllers/loader.controller");
+
+const router = express.Router();
+
+
+router.get("/", RenderLoadEpisodes)
+module.exports = router;
