+++ routes/loader/newloader.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const { RenderNewLoader } = require("../../controllers/loader.controller");
+
+const router = express.Router();
+
+router.post("/", RenderNewLoader)
+module.exports = router;
