+++ routes/manga/loadlatest.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const { RenderLatestManga } = require("../../controllers/manga.controller");
+const router = express.Router();
+
+router.get("/", RenderLatestManga);
+
+module.exports = router;
